# RepTracker.Api/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RepTracker.Api",
    "version": "v1"
  },
  "paths": {
    "/api/Country": {
      "get": {
        "tags": [
          "Country"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Country/{id}": {
      "get": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employee": {
      "get": {
        "tags": [
          "Employee"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Employee"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employee/{id}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobCategory": {
      "get": {
        "tags": [
          "JobCategory"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobCategory/{id}": {
      "get": {
        "tags": [
          "JobCategory"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Country": {
      "type": "object",
      "properties": {
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "MaritalStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "Gender": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "JobCategory": {
      "type": "object",
      "properties": {
        "jobCategoryId": {
          "format": "int32",
          "type": "integer"
        },
        "jobCategoryName": {
          "type": "string"
        }
      }
    },
    "Employee": {
      "required": [
        "email",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "employeeId": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "email": {
          "format": "email",
          "type": "string"
        },
        "street": {
          "type": "string"
        },
        "zip": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "country": {
          "$ref": "#/definitions/Country"
        },
        "phoneNumber": {
          "type": "string"
        },
        "smoker": {
          "type": "boolean"
        },
        "maritalStatus": {
          "$ref": "#/definitions/MaritalStatus"
        },
        "gender": {
          "$ref": "#/definitions/Gender"
        },
        "comment": {
          "maxLength": 1000,
          "minLength": 0,
          "type": "string"
        },
        "joinedDate": {
          "format": "date-time",
          "type": "string"
        },
        "exitDate": {
          "format": "date-time",
          "type": "string"
        },
        "jobCategoryId": {
          "format": "int32",
          "type": "integer"
        },
        "jobCategory": {
          "$ref": "#/definitions/JobCategory"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        }
      }
    }
  }
}